Create README.md
My Portfolio features a selection of projects that demonstrate my expertise in building scalable, user-friendly, and visually appealing applications. One of my recent projects is the MovieFilter highlights, some of the projects I have completed throughout my journey in IT. For instance, it shows my technical skills and creativity and demonstrates my passion for Information technology.